updated repo funcs with highlight funds
